add github release asset path root, tweak auth
usage:

  (let [bass (release:auth "vito/bass" "")]
    bass:assets/v0.1.1/bass.linux-amd64.tgz)

release:auth now runs 'gh auth login' which supports browser-based
interactive auth flow if no token is given. as a result the token is
stored in the image, but this should be safe as the image itself is
never exported or anything. assets may be downloaded and passed to other
thunks, but they would be mounted in without exposing the token that
fetched them.