re-introduce caching, use image digest for safety
previous attempts at this spooked me away from caching because swapping
out the base image broke the initial 'cp' call that seeds the cache with
the /nix/ content. the fear was that if they pushed a new image
everyone's caches would break.

but we can just take the image digest and put it in the cache mount dir!